requirements.txt
flask
flask_cors
pyOpenSSL
gunicorn
cryptography